Find header to print by cpu_type & various fixes

+++ srcs/slseek.c
@@ -6,22 +6,22 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/06 14:14:10 by jjaniec           #+#    #+#             */
-/*   Updated: 2019/12/06 14:14:30 by jjaniec          ###   ########.fr       */
+/*   Updated: 2019/12/13 19:18:54 by jjaniec          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <ft_nm.h>
 
-off_t		slseek(t_ft_nm_file *file, off_t offset, int whence)
+int		slseek(t_ft_nm_file *file, int offset, int whence)
 {
-	off_t	r;
+	int	r;
 
 	r = -1;
 	if (whence == SLSEEK_CUR)
 	{
 		if ((offset <= 0 && \
-			(size_t)offset < (size_t)((*file).seek_ptr - (*file).content)) || \
-			((size_t)offset + (size_t)((*file).seek_ptr - (*file).content) <= (*file).totsiz))
+			offset < ((*file).seek_ptr - (*file).content)) || \
+			(offset + ((*file).seek_ptr - (*file).content) <= (*file).totsiz))
 		{
 			(*file).seek_ptr += offset;
 			r = offset;
@@ -29,11 +29,12 @@ off_t		slseek(t_ft_nm_file *file, off_t offset, int whence)
 	}
 	else if (whence == SLSEEK_SET)
 	{
-		if (offset >= 0 && (*file).totsiz >= (size_t)offset)
+		if (offset >= 0 && (*file).totsiz >= offset)
 		{
 			(*file).seek_ptr = (*file).content + offset;
 			r = offset;
 		}
 	}
+	// dprintf(2, "Offset moved to: %ld\n", file->seek_ptr - file->content);
 	return (r);
 }
